feat: moved datatable from molecules to organisms
