+++ apps/backoffice-v2/src/lib/blocks/components/DataTableCell/DataTableCell.tsx
@@ -1,6 +1,6 @@
 import { FunctionComponent } from 'react';
 import { ExtractCellProps } from '@ballerine/blocks';
-import { DataTable } from '@/common/components/molecules/DataTable/DataTable';
+import { DataTable } from '@/common/components/organisms/DataTable/DataTable';
 
 export const DataTableCell: FunctionComponent<ExtractCellProps<'dataTable'>> = ({ value }) => (
   <DataTable {...value} />
